% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clopper_competitors}
\alias{clopper_competitors}
\title{Vowels that are likely to be confused in each dialect}
\format{A 18-row data frame with columns:
\describe{
  \item{Dialect}{Dialect region}
  \item{Vowel}{Vowel code}
  \item{Competitor}{Vowel code of a likely confusion for Vowel}
  \item{Description}{Reason for potential confusion}
}}
\usage{
clopper_competitors
}
\description{
Based on shifts described qualitatively in Clopper, Pisoni, and de Jong
(2005).
}
\seealso{
\code{\link{nsp_vows}}
}
\keyword{datasets}
