% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl.R
\name{KL.bvnorm}
\alias{KL.bvnorm}
\title{Calculate the KL divergence between two bivariate Gaussians}
\usage{
KL.bvnorm(mu1, sigma1, mu2, sigma2)
}
\arguments{
\item{mu1}{mean vector of first distribution}

\item{sigma1}{covariance matrix in vector (see covar2D) or matrix form}

\item{sigma2}{}
}
\description{
Calculate the KL divergence between two bivariate Gaussians
}
